fix whitespacebug in registration textfields, add x button to empty textfield
