+++ IBAN-Saver-App/Scenes/Registration Page/View/Components/RegTextFieldView.swift
@@ -21,3 +21,48 @@ struct RegTextFieldView: View {
         }
     }
 }
+
+
+struct CustomTextFieldView1: View {
+    @Binding var text: String
+    var placeholder: String
+    var isSecure: Bool = false
+    
+    var body: some View {
+            ZStack(alignment: .trailing) {
+                if isSecure {
+                    SecureField(placeholder, text: $text, prompt:
+                                    Text(placeholder).foregroundColor(.white.opacity(0.7)))
+                } else {
+                    TextField(placeholder, text: $text, prompt:
+                                Text(placeholder).foregroundColor(.white.opacity(0.7)))
+                }
+                
+                if !text.isEmpty {
+                    Button(action: {
+                        text = ""
+                    }) {
+                        Image(systemName: "xmark.circle.fill")
+                            .foregroundColor(.white.opacity(0.6))
+                            .padding(8)
+                    }
+                }
+            }
+            .textFieldStyle(CustomTextFieldStyle1())
+    }
+}
+
+
+// MARK: - TextfieldStyle
+struct CustomTextFieldStyle1: TextFieldStyle {
+    func _body(configuration: TextField<Self._Label>) -> some View {
+        configuration
+            .padding(16)
+            .background(
+                RoundedRectangle(cornerRadius: 10)
+                    .stroke(Color.white.opacity(0.4), lineWidth: 1)
+            )
+            .foregroundColor(.white)
+    }
+}
+
